hankyung_news_agent/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.44.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
pydantic>=2.8.0
openai>=1.40.0
langgraph>=0.2.0
langchain-core>=0.3.0
